refactor: CCTV 도메인을 shared/cctv 모듈로 추출 - yongin-platform의 CCTV 패키지를 shared/cctv 모듈로 이동 (com.pluxity.cctv) - CctvErrorCode 추출 (YonginErrorCode에서 CCTV 관련 6개 분리) - CctvProperties 추가 (즐겨찾기 최대 개수 yml 설정, 기본값 4) - MediaServerProperties를 shared/cctv로 이동 - Cctv 엔티티에 alt(고도) 컬럼 추가, name을 non-null로 변경 - 동기화 시 name을 streamName으로 초기화 - safers, yongin-platform 모두 shared:cctv 의존성 추가 - README에 shared/cctv 모듈 사용법 추가
