+++ apps/safers/src/main/resources/application-local.yml
@@ -53,3 +53,6 @@ logging:
 
 event:
   base-url: ${EVENT_BASE_URL:http://192.168.10.182:8001}
+
+media-server:
+  url: http://192.168.10.181:9997
