+++ apps/safers/src/main/resources/application.yml
@@ -1,4 +1,3 @@
-# deploy test
 spring:
   profiles:
     group:
