--- apps/yongin-platform/src/main/kotlin/com/pluxity/yongin/cctv/controller/CctvController.kt
@@ -1,95 +1,0 @@
-package com.pluxity.yongin.cctv.controller
-
-import com.pluxity.common.core.response.DataResponseBody
-import com.pluxity.common.core.response.ErrorResponseBody
-import com.pluxity.yongin.cctv.dto.CctvResponse
-import com.pluxity.yongin.cctv.dto.CctvUpdateRequest
-import com.pluxity.yongin.cctv.service.CctvService
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.media.Content
-import io.swagger.v3.oas.annotations.media.Schema
-import io.swagger.v3.oas.annotations.responses.ApiResponse
-import io.swagger.v3.oas.annotations.responses.ApiResponses
-import io.swagger.v3.oas.annotations.tags.Tag
-import jakarta.validation.Valid
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PatchMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/cctvs")
-@Tag(name = "CCTV Controller", description = "CCTV 관리 API")
-class CctvController(
-    private val service: CctvService,
-) {
-    @Operation(summary = "CCTV 동기화", description = "미디어서버에서 CCTV 경로 목록을 가져와 DB에 동기화합니다")
-    @ApiResponses(
-        value = [
-            ApiResponse(responseCode = "204", description = "동기화 성공"),
-            ApiResponse(
-                responseCode = "500",
-                description = "서버 오류",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ErrorResponseBody::class),
-                    ),
-                ],
-            ),
-        ],
-    )
-    @PostMapping("/sync")
-    fun sync(): ResponseEntity<Void> {
-        service.sync()
-        return ResponseEntity.noContent().build()
-    }
-
-    @Operation(summary = "CCTV 목록 조회", description = "CCTV 목록을 조회합니다 (즐겨찾기 우선, 이름순 정렬)")
-    @ApiResponses(
-        value = [
-            ApiResponse(responseCode = "200", description = "조회 성공"),
-            ApiResponse(
-                responseCode = "500",
-                description = "서버 오류",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ErrorResponseBody::class),
-                    ),
-                ],
-            ),
-        ],
-    )
-    @GetMapping
-    fun findAll(): ResponseEntity<DataResponseBody<List<CctvResponse>>> = ResponseEntity.ok(DataResponseBody(service.findAll()))
-
-    @Operation(summary = "CCTV 수정", description = "CCTV의 이름, 경도, 위도를 수정합니다")
-    @ApiResponses(
-        value = [
-            ApiResponse(responseCode = "204", description = "수정 성공"),
-            ApiResponse(
-                responseCode = "404",
-                description = "CCTV를 찾을 수 없음",
-                content = [
-                    Content(
-                        mediaType = "application/json",
-                        schema = Schema(implementation = ErrorResponseBody::class),
-                    ),
-                ],
-            ),
-        ],
-    )
-    @PatchMapping("/{id}")
-    fun update(
-        @PathVariable id: Long,
-        @RequestBody @Valid request: CctvUpdateRequest,
-    ): ResponseEntity<Void> {
-        service.update(id, request)
-        return ResponseEntity.noContent().build()
-    }
-}
