--- apps/yongin-platform/src/main/kotlin/com/pluxity/yongin/cctv/dto/CctvExternalResponse.kt
@@ -1,13 +1,0 @@
-package com.pluxity.yongin.cctv.dto
-
-data class MediaServerPathListResponse(
-    val itemCount: Int,
-    val pageCount: Int,
-    val items: List<MediaServerPathItem>,
-)
-
-data class MediaServerPathItem(
-    val name: String,
-    val confName: String,
-    val ready: Boolean,
-)
