--- apps/yongin-platform/src/main/kotlin/com/pluxity/yongin/cctv/repository/CctvRepository.kt
@@ -1,6 +1,0 @@
-package com.pluxity.yongin.cctv.repository
-
-import com.pluxity.yongin.cctv.entity.Cctv
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CctvRepository : JpaRepository<Cctv, Long>
