+++ apps/yongin-platform/src/main/kotlin/com/pluxity/yongin/global/constant/YonginErrorCode.kt
@@ -20,12 +20,6 @@ enum class YonginErrorCode(
     INVALID_NOTICE_DATE_RANGE(HttpStatus.BAD_REQUEST, "시작일이 종료일보다 이후일 수 없습니다."),
     NOT_FOUND_ATTENDANCE(HttpStatus.NOT_FOUND, "ID가 %s인 출역현황을 찾을 수 없습니다."),
     NOT_FOUND_OBSERVATION(HttpStatus.NOT_FOUND, "ID가 %s인 드론 관측 데이터를 찾을 수 없습니다."),
-    NOT_FOUND_CCTV(HttpStatus.NOT_FOUND, "ID가 %s인 CCTV를 찾을 수 없습니다."),
-    MEDIA_SERVER_URL_NOT_CONFIGURED(HttpStatus.INTERNAL_SERVER_ERROR, "미디어서버 URL이 설정되지 않았습니다."),
-    EXCEED_FAVORITE_LIMIT(HttpStatus.BAD_REQUEST, "즐겨찾기는 최대 4개까지 추가할 수 있습니다."),
-    ALREADY_FAVORITE(HttpStatus.BAD_REQUEST, "스트림명이 %s인 CCTV는 이미 즐겨찾기되어 있습니다."),
-    INVALID_FAVORITE_ORDER_COUNT(HttpStatus.BAD_REQUEST, "즐겨찾기 순서 변경은 전체 즐겨찾기를 대상으로 해야 합니다."),
-    NOT_FOUND_CCTV_FAVORITE(HttpStatus.NOT_FOUND, "ID가 %s인 CCTV 즐겨찾기를 찾을 수 없습니다."),
     NOT_FOUND_SAFETY_EQUIPMENT(HttpStatus.NOT_FOUND, "ID가 %s인 안전장비를 찾을 수 없습니다."),
     ;
