--- apps/yongin-platform/src/main/kotlin/com/pluxity/yongin/global/properties/MediaServerProperties.kt
@@ -1,11 +1,0 @@
-package com.pluxity.yongin.global.properties
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.bind.ConstructorBinding
-
-@ConfigurationProperties(prefix = "media-server")
-data class MediaServerProperties
-    @ConstructorBinding
-    constructor(
-        val url: String,
-    )
