+++ settings.gradle.kts
@@ -7,6 +7,9 @@ include(":common:file")
 include(":common:messaging")
 include(":common:test-support")
 
+// ── Shared Modules ──
+include(":shared:cctv")
+
 // ── Application Modules ──
 include(":apps:safers")
 include(":apps:yongin-platform")
