+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/config/CctvConfig.kt
@@ -1,0 +1,8 @@
+package com.pluxity.cctv.config
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableConfigurationProperties(CctvProperties::class, MediaServerProperties::class)
+class CctvConfig
