+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/config/CctvProperties.kt
@@ -1,0 +1,8 @@
+package com.pluxity.cctv.config
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "cctv")
+data class CctvProperties(
+    val maxFavoriteCount: Int = 4,
+)
