+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/config/MediaServerProperties.kt
@@ -1,0 +1,8 @@
+package com.pluxity.cctv.config
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "media-server")
+data class MediaServerProperties(
+    val url: String = "",
+)
