+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/dto/CctvExternalResponse.kt
@@ -1,0 +1,13 @@
+package com.pluxity.cctv.dto
+
+data class MediaServerPathListResponse(
+    val itemCount: Int,
+    val pageCount: Int,
+    val items: List<MediaServerPathItem>,
+)
+
+data class MediaServerPathItem(
+    val name: String,
+    val confName: String,
+    val ready: Boolean,
+)
