+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/entity/Cctv.kt
@@ -1,0 +1,33 @@
+package com.pluxity.cctv.entity
+
+import com.pluxity.common.core.entity.IdentityIdEntity
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.Table
+
+@Entity
+@Table(name = "cctv_stream")
+class Cctv(
+    @Column(name = "stream_name", nullable = false, unique = true)
+    val streamName: String,
+    @Column(name = "name", nullable = false)
+    var name: String,
+    @Column(name = "lon")
+    var lon: Double? = null,
+    @Column(name = "lat")
+    var lat: Double? = null,
+    @Column(name = "alt")
+    var alt: Double? = null,
+) : IdentityIdEntity() {
+    fun update(
+        name: String,
+        lon: Double?,
+        lat: Double?,
+        alt: Double?,
+    ) {
+        this.name = name
+        this.lon = lon
+        this.lat = lat
+        this.alt = alt
+    }
+}
