+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/repository/CctvFavoriteRepository.kt
@@ -1,0 +1,10 @@
+package com.pluxity.cctv.repository
+
+import com.pluxity.cctv.entity.CctvFavorite
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CctvFavoriteRepository : JpaRepository<CctvFavorite, Long> {
+    fun findAllByOrderByDisplayOrderAsc(): List<CctvFavorite>
+
+    fun existsByStreamName(streamName: String): Boolean
+}
