+++ shared/cctv/src/main/kotlin/com/pluxity/cctv/repository/CctvRepository.kt
@@ -1,0 +1,6 @@
+package com.pluxity.cctv.repository
+
+import com.pluxity.cctv.entity.Cctv
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CctvRepository : JpaRepository<Cctv, Long>
